test(node): remove duplicate test cases
