--- packages/rolldown/tests/fixtures/plugin/normalized-option/main.js
@@ -1,1 +1,0 @@
-console.log()
